requirements.txt
# AutoU Email App - Dependências Otimizadas para Deploy
flask==3.0.3
werkzeug==3.0.3
gunicorn==21.2.0
pdfminer.six==20231228
PyPDF2>=3.0.0